Moved various generalized blotter test functions and mocks into separate files.
