+++ src/client/src/applicationServices.mock.ts
@@ -1,0 +1,13 @@
+import { ApplicationDependencies } from 'applicationServices'
+
+const MockApplicationDependencies = jest.fn<ApplicationDependencies>(
+  (loadBalancedServiceStub?, referenceDataService?, platform?, limitChecker?, serviceStatus$?, connection$?) => ({
+    referenceDataService,
+    platform,
+    limitChecker,
+    loadBalancedServiceStub,
+    serviceStatus$,
+    connection$,
+  }),
+)
+export default MockApplicationDependencies
